backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
twilio
python-dotenv
httpx 